Enemy navmesh and wave enemy spawn system
Added enemy NavMesh system for the chasing building (Main building)
Added a system for spawning enemies in waves
Added events to check map creation, bake navmesh and GameManager.

There is a problem with the enemy spawning system, in particular, the spawn points are filled in the list only after restarting the program (code change, restarting unity). At next launches of the game, you can see in the debug that the list is filled with spawn points, but in the editor the list is empty, and, as a consequence, the spawning system does not know where to spawn enemies